+++ src/iteration_utilities/_utils.py
@@ -4,8 +4,11 @@
 import platform
 import sys
 
-__all__ = ['GE_PY38', 'IS_CPYTHON', 'IS_PYPY', 'USES_VECTORCALL', '_default']
+__all__ = ['GE_PY36', 'GE_PY38', 'IS_CPYTHON', 'IS_PYPY', 'USES_VECTORCALL', '_default']
 
+
+GE_PY36 = sys.version_info.major > 3 or (sys.version_info.major == 3 and
+                                         sys.version_info.minor >= 6)
 GE_PY38 = sys.version_info.major > 3 or (sys.version_info.major == 3 and
                                          sys.version_info.minor >= 8)
 IS_PYPY = platform.python_implementation() == 'PyPy'